requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pgvector==0.3.5
openai>=2.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
pydantic[email]>=2.0.0
